Add the --akaros extension
Which right now further makes error messages that are useful
and don't embed binary data in what should be a string.

If --akaros is set, then Rerror strings look like this:
XXXX message

The XXXX is the errno encoded in HEX as text. As it was in 1998
in the original v9fs :-)

Signed-off-by: Ronald G. Minnich <rminnich@google.com>